sync challenges

+++ challenges/easy/1_vector_add/starter/starter.cu
@@ -1,9 +1,9 @@
 #include <cuda_runtime.h>
 
-__global__ void vector_add(const float* A, const float* B, float* C, size_t N) {}
+__global__ void vector_add(const float* A, const float* B, float* C, int N) {}
 
 // A, B, C are device pointers (i.e. pointers to memory on the GPU)
-extern "C" void solve(const float* A, const float* B, float* C, size_t N) {
+extern "C" void solve(const float* A, const float* B, float* C, int N) {
     int threadsPerBlock = 256;
     int blocksPerGrid = (N + threadsPerBlock - 1) / threadsPerBlock;
